Solutions for 8 and 9(though only part 1 of nine so far)
Found an error on problem 9. TLDR is that the example and final solution are slightly
wrong for the first chunk. I had to add in a hack to use denseMap-1 for example to pass,
but using that causes the real version to fail.

This may come down to how I am doing the backfill, but I think it may be an actual problem