part 1 solved, not sure why my solution for part 2 is breaking down
Initial thought was that I was exiting too early, not catching delta
at bounds. Increased range and that was A problem

Second thought, problem with incrementing lastPoint when handling
disjoint sides. Added myTest with some tricky geometry but did
not encounter any errors with the processing

Moving on to day 13